Add tests for invites

+++ lib/ksomnia/invite.ex
@@ -24,11 +24,25 @@ defmodule Ksomnia.Invite do
     invite
     |> cast(attrs, [:email])
     |> validate_required([:email])
+    |> unique_constraint([:email, :team_id])
   end
 
   def new(team_id, attrs) do
     %Invite{team_id: team_id}
     |> changeset(attrs)
+    |> ensure_not_member
+  end
+
+  def ensure_not_member(changeset) do
+    email = changeset.changes.email
+    team_id = changeset.data.team_id
+
+    if TeamUser.is_member(team_id, email) do
+      changeset
+      |> add_error(team_id, "the user is already a team member")
+    else
+      changeset
+    end
   end
 
   def create(team_id, attrs) do
@@ -50,8 +64,12 @@ defmodule Ksomnia.Invite do
     Repo.delete(invite)
   end
 
-  def reject(invite) do
-    Repo.delete(invite)
+  def reject(invite_id, invitee) do
+    if invite = Repo.get_by(Invite, email: invitee.email, id: invite_id) do
+      Repo.delete(invite)
+    else
+      :error
+    end
   end
 
   def accept(invite_id, invitee) when is_binary(invite_id) do
